requirements.txt
streamlit==1.35.0
tensorflow==2.10.0
numpy==1.23.5
Pillow==9.4.0
matplotlib==3.4.2
seaborn==0.11.1
pandas==1.2.4
protobuf==3.20.*
